+++ docs/docs/.vitepress/config.ts
@@ -1,0 +1,38 @@
+import {defineConfig} from 'vitepress'
+
+export default defineConfig({
+  title: 'Vite Plugin',
+  description: 'Documentation for the Vite plugin',
+  base: '/docs/vite/',
+  lang: 'en-US',
+  head: [
+    ['meta', {content: 'https://github.com/nystudio107', property: 'og:see_also',}],
+    ['meta', {content: 'https://twitter.com/nystudio107', property: 'og:see_also',}],
+    ['meta', {content: 'https://youtube.com/nystudio107', property: 'og:see_also',}],
+    ['meta', {content: 'https://www.facebook.com/newyorkstudio107', property: 'og:see_also',}],
+  ],
+  themeConfig: {
+    socialLinks: [
+      {icon: 'github', link: 'https://github.com/nystudio107'},
+      {icon: 'twitter', link: 'https://twitter.com/nystudio107'},
+    ],
+    logo: '/img/plugin-logo.svg',
+    editLink: {
+      pattern: 'https://github.com/nystudio107/craft-vite/edit/develop/docs/docs/:path',
+      text: 'Edit this page on GitHub'
+    },
+    algolia: {
+      appId: 'AE3HRUJFEW',
+      apiKey: 'c5dcc2be096fff3a4714c3a877a056fa',
+      indexName: 'vite'
+    },
+    lastUpdatedText: 'Last Updated',
+    sidebar: [],
+    nav: [
+      {text: 'Home', link: 'https://nystudio107.com/plugins/vite'},
+      {text: 'Store', link: 'https://plugins.craftcms.com/vite'},
+      {text: 'Changelog', link: 'https://nystudio107.com/plugins/vite/changelog'},
+      {text: 'Issues', link: 'https://github.com/nystudio107/craft-vite/issues'},
+    ]
+  },
+});