+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## 4.0.4 - 2023.01.01
+### Changed
+* Move to using `ServicesTrait` and add getter methods for services
+* Update the docs to use Vitepress `^1.0.0-alpha.29`
+
+### Fixed
+* Fixed an issue where `craft.vite.includeCriticalCssTags()` would fail if you had `template` in the file path ([#45](https://github.com/nystudio107/craft-vite/issues/45))
+
 ## 4.0.3 - 2022.09.09
 ### Added
 * Added support for detecting dev-mode in Craft CMS v4 by changing `App::env('ENVIRONMENT') === 'dev'`